consolidate tests into test_torch_trace_coverage.py, removed noqa from inject_roctx_loader.py
